ChangeList: The segment argument used with _dos_allocmem(), _dos_freemem() and _dos_setblock() is now unsigned int instead of unsigned short. This change was made for compatibility with other compilers.
Note: Tell 32-bit DOS users that they need to recompile, and that
they shouldn't be using these functions in the first place.

[git-p4: depot-paths = "//depot/openwatcom/": change = 31993]

+++ bld/clib/_dos/c/_dostest.c
@@ -151,7 +151,7 @@ void TestMemoryOperations( void )
     void                *segment;
     char                *cptr;
     #else
-    unsigned short      segment;
+    unsigned            segment;
     char __far          *cptr;
     #endif
     unsigned            ctr;
